chore: rename swap-specific names
